authentication check

+++ website/fileExplorer.py
@@ -34,26 +34,35 @@ def file_explorer():
     if request.method == "POST":
         file = request.files["file"]
         if file:
-            existance = Files.query.filter_by(owner = user.id).all()
-            upload = Files(filename=file.filename, owner = current_user.id)
+            existance = Files.query.filter_by(user_id = user.id).all()
+            
             for exist in existance:
                 if exist.filename == file.filename:
                     duplicated = True
             
             if not duplicated:
-                db.session.add(upload)
-                db.session.commit()
-
                 path = f'C:\ISTEC\PROJETO FINAL\TESTES\webserver\\files\{user.id}\{file.filename}'
                 file.save(path)
+                
+                file_stats = os.stat(path)
+                filesize = str(round(file_stats.st_size / (1024 * 1024), 2))
+                
+                
+                
+                upload = Files(filename=file.filename, username = user.username, user_image = current_user.profile_pic, user_id = current_user.id, size = filesize)
+                
+                db.session.add(upload)
+                db.session.commit()
                 flash("O seu ficheiro foi carregado com sucesso", "success")
+                
+                
             else:
                 flash("Já existe um ficheiro com esse nome, apague-o ou mude o nome do mesmo", "danger")
         else:
             flash("Insira um ficheiro", "warning")
+    
         
-    messages = get_flashed_messages(with_categories=True, category_filter=["success", "info", "warning", "danger"])
-    return render_template("file_explorer.html", client=user, messages=messages)
+    return render_template("file_explorer.html", client=user)
 
 
 @fileExplorer.route('/file-explorer/download/<path:fileId>', methods=["GET","POST"])
@@ -61,8 +70,12 @@ def file_explorer():
 def download(fileId):
     user = User.query.get(current_user.id)
     file = Files.query.get(fileId)
-    path = f'C:\ISTEC\PROJETO FINAL\TESTES\webserver\\files\{user.id}\{file.filename}'
-    return send_file(path, as_attachment=True)
+    
+    if user.id == file.user_id:
+        path = f'C:\ISTEC\PROJETO FINAL\TESTES\webserver\\files\{user.id}\{file.filename}'
+        return send_file(path, as_attachment=True)
+    else:
+        return flash("You don't have permission to access that file!", "danger")
         
 
 @fileExplorer.route('/file-explorer/delete-file/<path:fileId>', methods=["GET","POST"])
@@ -72,24 +85,13 @@ def deleteFile(fileId):
 
     user = User.query.get(current_user.id)
     file = Files.query.get(fileId)
-    path = f'C:\ISTEC\PROJETO FINAL\TESTES\webserver\\files\{user.id}\{file.filename}'
-    os.remove(path)
-    db.session.delete(file)
-    db.session.commit()
-    flash(f"O ficheiro {file.filename} foi apagado com sucesso.", "success")
-    return redirect(url_for('fileExplorer.file_explorer'))
-
-
-
-
-
-# nao é mais preciso porque ja esta a funcioar no def file_explorer
-# @views.route('/file-explorer/upload', methods=['POST'])
-# def upload_file():
-#     print("ta no upload")
-#     user = User.query.get(current_user.id)
-#     if request.method == 'POST':
-#       f = request.files['file']
-#       path = f'C:\ISTEC\PROJETO FINAL\TESTES\webserver\\files\{user.id}\{secure_filename(f.filename)}'
-#       f.save(path)
-#       return redirect('/file-explorer')
+    
+    if user.id == file.user_id:
+        path = f'C:\ISTEC\PROJETO FINAL\TESTES\webserver\\files\{user.id}\{file.filename}'
+        os.remove(path)
+        db.session.delete(file)
+        db.session.commit()
+        flash(rf"O ficheiro <b>{file.filename}</b> foi apagado com sucesso.", "success")
+        return redirect(url_for('fileExplorer.file_explorer'))
+    else:
+        return flash("You don't have permission to access that file!", "danger")
